feat: добавлены типы для order-details-slice
